+++ packages/charts/src/components/LineChart/Placeholder.tsx
@@ -5,18 +5,21 @@ import ContentLoader from 'react-content-loader';
 export const LineChartPlaceholder = (props) => {
   return (
     <ContentLoader
-      height={300}
-      width={300}
+      viewBox="0 0 165 145"
+      preserveAspectRatio="xMidYMid meet"
+      width="100%"
+      height="100%"
       speed={2}
       backgroundColor={ThemingParameters.sapContent_ImagePlaceholderBackground}
       foregroundColor={ThemingParameters.sapContent_ImagePlaceholderForegroundColor}
       backgroundOpacity={ThemingParameters.sapContent_DisabledOpacity as any}
       {...props}
     >
+      <rect x="10" y="135" width="135" height="1" />
       <path
-        d="M30 260 L60 190 L90 150 L120 90 L150 140 L180 110 L210 80 L240 150 L270 130 L270 135 L240 155 L210 85 L180 115 L150 145 L120 95 L90 155 L60 195 L30 265 Z"
-        fill="transparent"
-        strokeWidth="5"
+        d="M20 135 L40 80 L60 70 L80 30 L100 50 L120 65 L 135 100 L120 67 L100 52 L80 32 L60 72 L40 82 L20 137 Z"
+        fill="none"
+        strokeWidth="2"
       />
     </ContentLoader>
   );