+++ packages/charts/src/lib/components/ChartDataLabel.ts
@@ -1,0 +1,3 @@
+import { ChartDataLabel } from '../../internal/ChartDataLabel';
+
+export { ChartDataLabel };